+++ Packages/com.thenathannator.plasticband/PlasticBand/Devices/FiveFretGuitar/RockBandGuitar/SantrollerXInputRockBandGuitar.cs
@@ -9,36 +9,37 @@ using UnityEngine.InputSystem.Utilities;
 using UnityEngine.InputSystem.XInput;
 
 // PlasticBand reference doc:
-// https://github.com/TheNathannator/PlasticBand/blob/main/Docs/Instruments/5-Fret%20Guitar/Guitar%20Hero/Xbox%20360.md
+// https://github.com/TheNathannator/PlasticBand/blob/main/Docs/Instruments/5-Fret%20Guitar/Rock%20Band/Xbox%20360.md
+
 namespace PlasticBand.Devices
 {
     /// <summary>
-    /// A Santroller XInput Guitar Hero Guitar.
+    /// A Santroller XInput Rock Band Guitar.
     /// </summary>
-    [InputControlLayout(stateType = typeof(XInputGuitarAlternateState), displayName = "Santroller XInput Guitar Hero Guitar")]
-    public class SantrollerXInputGuitarAlternate : XInputGuitarAlternate
+    [InputControlLayout(stateType = typeof(XInputRockBandGuitarState), displayName = "Santroller XInput Rock Band Guitar")]
+    public class SantrollerXInputRockBandGuitar : XInputRockBandGuitar
     {
         /// <summary>
-        /// The current <see cref="SantrollerXInputGuitarAlternate"/>.
+        /// The current <see cref="SantrollerXInputRockBandGuitar"/>.
         /// </summary>
-        public static new SantrollerXInputGuitarAlternate current { get; private set; }
+        public static new SantrollerXInputRockBandGuitar current { get; private set; }
 
         /// <summary>
-        /// A collection of all <see cref="SantrollerXInputGuitarAlternate"/>s currently connected to the system.
+        /// A collection of all <see cref="SantrollerXInputRockBandGuitar"/>s currently connected to the system.
         /// </summary>
-        public new static IReadOnlyList<SantrollerXInputGuitarAlternate> all => s_AllDevices;
-        private static readonly List<SantrollerXInputGuitarAlternate> s_AllDevices = new List<SantrollerXInputGuitarAlternate>();
+        public new static IReadOnlyList<SantrollerXInputRockBandGuitar> all => s_AllDevices;
+        private static readonly List<SantrollerXInputRockBandGuitar> s_AllDevices = new List<SantrollerXInputRockBandGuitar>();
 
         /// <summary>
-        /// Registers <see cref="SantrollerXInputGuitarAlternate"/> to the input system.
+        /// Registers <see cref="SantrollerXInputRockBandGuitar"/> to the input system.
         /// </summary>
         internal new static void Initialize()
         {
-            SantrollerLayoutFinder.Register<SantrollerXInputGuitarAlternate>(XInputController.DeviceSubType.GuitarAlternate);
+            SantrollerLayoutFinder.Register<SantrollerXInputRockBandGuitar>(XInputController.DeviceSubType.Guitar);
         }
 
         /// <summary>
-        /// Sets this device as the current <see cref="SantrollerXInputGuitarAlternate"/>.
+        /// Sets this device as the current <see cref="SantrollerXInputRockBandGuitar"/>.
         /// </summary>
         public override void MakeCurrent()
         {